simulator/requirements.txt
psycopg2-binary>=2.9,<3
pyyaml>=6.0,<7
